Rename OT_prac2.py to TP_Cost_Minimum.py
The Cost Minimum Method , alternatively known as Least Cost method or Matrix Minima Method, is used to find an optimum solution to a transportation problem. The objective of a transportation problem is to minimize the cost of distributing a product from a number of sources to destinations. There are several methods to obtain a minimum transportation cost, however non guarantee an optimum solution. Hence Modified Distribution method is used to test the feasibility of the solution obtained